Add yet another route to jails: an interface route for the real control net. Previously, if you accessed a phys node by its real IP address (155.101.132.N) from inside a jail, it used the default route going to the router and back. Now it will put it out directly on the control net.
The main reason for this is so that minibed vnodes can reach their boss and
ops nodes.  Vnodes in mini don't have a default route that works, so we needed
a way to get to boss/ops.

Note that gated will remove this route when it starts up.  Thus, vnodes will
revert to using the default route to get to the real control net.  On mainbed,
this will still work (hop through the router as before).  On mini, it won't
work at all.

Moral: don't use session routing and vnodes on mini.