Commit the stuff necessary to copy out new plab rootballs, versions of which had been hanging around in my home directory for a while.
There are a few new things in plab/etc/netbed_files that set up a
directory of the same name in @prefix@. This will get rsync'ed with
netbed_files/ on each planetlab node.
log/  - just needs to exist for the httpd server
sbin/ - contains thttpd, and scripts to manipulate it
www/  - the directory served by thttpd. Contains symlinks to the 'real'
        location of the rootballs (etc/plab)

I've committed a binary of thttpd - this is simply because it'd be a
PITA to compile a Linux binary for every devel tree, etc.

PLAB_ROOTBALL has now become a configure options. The idea is that we
will keep the latest version number in configure.in, but you can
override it in your defs
file. This way, we don't have to update every defs file when there's a new
version, but people can still play around with their own version if they want.

The two scripts that interact with the plab nodes skip ones that are
down. They ssh in as 'utah1', meaning that one of us who has access to
that account needs to run them, so that they can have access to our
keys. We can put boss's public key (or something) out there to remove
this requirement.

plabdist runs an rsync between @prefix@/etc/plab/netbed_files and a
file of the same name on the planetlab nodes. It's intended to be run
from the main install tree - the local rsync directory is not normally
set up in devel trees. It runs in parallel, but is limited to 4 to
avoid beating up boss too much. Takes about 1:40 with the current set
of plab nodes (took > 10 minutes doing one at a time).

plabhttpd (re)starts the mini web server on all plab nodes