Bring back the Alternative PHP cache. The apc.so module is installed in the php extensions directory, and the following two lines were added to /usr/local/etc/php.ini:
	extension_dir = /usr/local/lib/php/20020429/
	extension="apc.so"

The new version is a little odd in that it is more difficult to clear
the cache, which we must do when we install new pages. Rather then
jump through too many hoops, I added a simple clearapc.php page, which
does nothing but make the call from within the webserver. I have
hooked this into the make install target using wget. If you care to
see what is currently in the cache, you can load this page:

	https://www.emulab.net/apc.php

which gives a non-pretty print dump of the cache. To clear the cache
from your web browser:

	https://www.emulab.net/clearapc.php