A few fixes for 'loopback' links in the virtual topology (links that connect a vnode to itself):
Move up the point at which we consider a vnode to be assigned - we
want to do this before we look for link resolutions, or the 'other
end' of a loopback link appears to be unassigned, which is not true.

Add a mecahism to watch out for scoring lookpback links twice. The
edge iterator for an undirected graph sees each vlink twice (there
are two 'out' edges from the vnode). So, we have to avoid scoring it
the second time.

Note that we still only allow loopback links if they are 'trivial_ok'.