Added some new display abilities to menu.php3 .
PAGEHEADER() now takes a vew array, which can turn on and off display
of things like the sidebar and banner.

In the future, some of this information will likely be grabbed from
the user somehow (new and/or plab users may get a different view than
others).

Also added the functions to draw a 'topbar' of options that goes
across the top of the page, rather than down the left side. A simple
topbar for planetlab nodes is included.