Some support for gigabit interfaces. The biggest change is that we're more liberal about which links don't need delays. Before, we were putting delays on anything that wasn't 100Mbps.
The other change is that for delay nodes, we set virtual link
bandwidths by rounding up to the next fastest speed of interface,
instead of hardwiring 100Mbps. Similarly, we set the speed of
physical delay interfaces to the card's 'native' speed instead of
hardwiring 100Mbps.

These changes work for our nodes and links, but are still not totally
general. Really, what needs to happen is that assign needs to be the
one to insert delay nodes, because it's the only place where we really
know which interfaces of what speeds are available on which nodes.
But, that is a very big project, since both assign wrapper and assign
need to be changed.