New script.
Grant permission to use types/class of nodes. If given a specific
"type", then grant permission to use just that type. If its a class, then
permission to use all of the types in that class, the class itself, and any
aux nodetypes for the type/class (node_types_auxtypes table).

For example:

	wap grantnodetype -p testbed pc2000
	wap grantnodetype -p testbed pcvm
	wap grantnodetype -p testbed pcplab