Move setting the node permission table for a project from the web interface to the backend. mkproj now looks at the pcremote_ok set and makes the proper calls to grantnodetype. This reduces the amount of hardwired goo in the web interface.
Still, there is a bit of hardwired stuff in mkproj. At present we do
not form a relationship between a phys node type and the types we
assign to the virtual nodes. Thats is, nothing says that a pcplabphys
implies the right to use pcplabinet, etc. With only 3 remote phys
types, I just hardwired it into mkproj calling grantnodetype with type
pcplab (the class for the virtnodes) for pcplabphys. Same for pcron
and pcwa, (both get pcvwa). Ultimately we need a better type system.
In general the type system is pretty screwy.