Add 'local' features and desires. As opposed to 'global' fds, these effect a single node, but are different from 'class' fds in that they are stateful - ie. the socre you get depends on all other vnodes currently mapped to the pnodes. These are identified by prefixing them with a '?'. ('*' for global, mnemoic 'matches all'; '?' for local, mnemonic 'matches one')
The currently implemented local fd is the additive one, or '?+'. We
add the desire weight for all vnodes on the pnode together, and if
this number is greater than the pnode's weight for the feature, a
violation is flagged.

The idea here is that we can give a pnode a feature like: ?+memory:256
and vnodes desires like ?+memory:16 or ?+memory:64 to pack vnodes
onto pnodes in a manner that appropriately reflects their resource
usage.