A couple of things:
1) Added PLAB_SLICEPREFIX so that we can separately instantiate plab slices
from mini, or elsewhere.  On the mainbed, its set to "emulab".  On mini, its
set to "emulab_mini".  The "emulab" part has to exist first so that the new
plab node manager doesn't nuke our dslice slivers.

2) Fixed up Plab.getFree() so that it doesn't try to add the same IP twice
to the DB if a new one is found, and listed more than once.