Commit to usher in the new PLC regime.  Added a config variable to vnode_setup for the timeout on waiting for child processes.  I've set it to 10 minutes since all ancillary setup programs have their own time bounds (I think - the plab ones do anyway).
The function of plabmonitord has changed slightly.  Instead of setting
up and tearing down vnodes, its job is to just setup the emulab management
sliver on plab nodes in hwdown.  Once the vserver comes up and reports isalive,
it moves the node out of hwdown.  Currently, it first tries to tear down the
vserver before reinstantiating it.  In the future, we could get fancier and
try interacting with the service sliver directly before simply tearing it down.

All new plab nodes now start life in hwdown, and must be summoned forth
into production by plabmonitord.

This commit does NOT include support for the node-local httpd.  That will
come soon.