Yet another bugfix + code to call a function that sends NSESWAP event when it cannot keep up with real-time.
bug: This affected encapsulated simulator packets that had to cross
multiple physical nodes before arriving at the destination simulator
traffic agent. This bug didnt affect live packets from traffic sources
on real PCs.

The NSESWAP event is now sent via the tevc command. The nse scheduler
waits for the slop factor (diff between clock and event dispatch time
that exceeds a threshold) to be crossed multiple times in a second
before sending the NSESWAP event. Currently 5 times in 1 second.
However, this needs more careful thought and will get modified later.
When is it really necessary to declare that an nse is overloaded?
i.e. what is the right slop factor? How many times can we tolerate
that the slop factor is exceeded to ensure end-to-end performance
is within a certain percentage of the expected?