Checkpoint commit. I've added the static-linked program "ltevent" to send and wait for linktest events, and integrated it into the perl scripts.
At this point I'm working on small potatoes such as the rc.setup script
to invoke linktest daemon and nailing down where to put platform-specific
stuff such as the ns-linktest build and the customized tb-compat.

Also am testing linktest in cases where I've deliberately inserted an error
to make sure it catches it.