Redo run_linktest.pl so that it is okay to run on boss/ops. Note that I am not using paperbag to run it on ops, but just installing the two programs that are needed. Eventually should paperbag it.
This same script runs on both boss/ops and on experimental nodes to
fire off the linktest daemon on the nodes and wait. This is icky as
some crufty stuff has to be done so that it will run in both
environments. ltevent is of course not needed on boss, but do not want
to bother it now since it *is* needed on the nodes.

To run it on boss script needed severe cleanup and taint checking.
Also added the use of event keyfile, which is currently optional, but
needs to be mandatory once the images are updated.

The current problem is that run_linktest can hang waiting for the
clients to finish?. We can use quick mode in the experiment setup
path, but maybe we need a cancel operation?