New script: assign_prepass
Acts as a transparent filter for assign - parses the top file given to
assign, coarsens the graph in it, then feeds that to assign.
Translates assign's output back into context of the original graph,
and spits that out for parsing.

Right now, only includes a single coarsening algorithm - conglomerates
the nodes in leaf lans. Hopefully, it will soon include other
algorithms, such as running METIS to partition the graph.

Still needs a fair amount of work to be general - only works on
certain types of topologies right now.

Not called yet, but that will be trivial - just need to change
assign_wrapper to call assign_prepass instead of assign.