Medium size cleanup of the client side code. The main impetus behind this was to add soft reconfig support so that nodes could be reconfigured without having to reboot them. This appears to work, and has been tested with jails getting moved around. I've also tested the new code on the MFS, but still no testing has been done on PLAB nodes.
The main change is that most of the code moved out of libsetup.pm, and
was split into constituent rc scripts, each of which does its own
thing, including cleaning up and preparing for making an image. Most
of that central knowledge has been moved out into the scripts. Still more
to do but this was a good start.