Bug Fix: Do not export the homedirs of users who are "webonly" or who are not "active" (perhaps frozen).
Also, as a sanity check and to prevent mountd from barfing, make sure
that any directory we try to export, actually exists. This is done via
the local mount points on boss, and is perhaps not such a great idea
cause it is via NFS. Keep an eye out for problems.