Kirk takes the weed whacker to the plab code.  This is the first pass result. I'll come along for a closer cut in the future.
* Modularized the plab communications 'adaptor' interface and moved the
  dslice- and PLC-specific code into their own modules.

* Wrote an API definition README

* Separated out generic routines from libplab into their own library modules
  (libtestbed.py and libdb.py)

Functionally, not much has changed - this was just a massive re-org with some
other cleanup.  Should be much easier to code up new PLAB interfaces as the
plab folks flail around in their attempt to standardize on something.

XXX: may want to re-think where the generic library modules should go.  If
more python code enters Elab, we'll probably want to move 'em to more standard
locations.

This isn't the end of the cleanup - I would eventually like to go back and
rethink the class structures, beef up the comments, and extend the API.