Clean up the way we handle vnodes and unroutable IP addresses - there isn't really a good reason for keeping them seperate. All we really care about is putting the unroutable addresses in one file and the routable ones in another.
So, instead of generating three forward zone files, we only have to
generate two. They are not as pretty as the ones named_setup used to
generate, but the code to make them is much, much cleaner and more
comprehensible.

Also makes named_setup about twice as fast - 2s vs 4s. Still too
slow, but better!